refactor the live editor
